Removed base64 and added logger as an explicit add_dependency
Starting with Ruby 3.4, base64 is no longer a bundled gem, and now Thrift does not load in modern Ruby versions:

    An error occurred while loading ./spec/types_spec.rb.

    Failure/Error: require 'base64'

    LoadError:
      cannot load such file -- base64
    # ./lib/thrift/protocol/json_protocol.rb:21:in '<top (required)>'
    # ./lib/thrift.rb:45:in '<top (required)>'
    # ./spec/spec_helper.rb:30:in '<top (required)>'
    # ./spec/types_spec.rb:20:in '<top (required)>'

Ruby already has ability to serialize and deserialize base64 without requiring the base64 gem,
which is a thin wrapper for syntactic sugar.

Additionally, the code throws a warning at the moment, which will become an error in Ruby 3.5:

    /code/lib/thrift/processor.rb:20: warning: logger was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
    You can add logger to your Gemfile or gemspec to silence this warning.

Added logger as an explicit dependency to avoid this warning.